feat: display downloaded data in grid format with reusable components

+++ Crypto/Models/Coin.swift
@@ -11,7 +11,7 @@ struct Coin: Identifiable, Codable {
     let id, symbol, name: String
     let image: String
     let currentPrice: Double
-    let marketCap, marketCapRank, fullyDilutedValuation: Int?
+    let marketCap, marketCapRank, fullyDilutedValuation: Double?
     let totalVolume, high24H, low24H: Double?
     let priceChange24H, priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
